Module Tezos_client_sapling_010_PtGRANAD

module Client_sapling_commands : sig ... end

module Context : sig ... end

This module allows the creation Sapling transactions: shield, unshield and transfer. Because Sapling uses an UTXO model, it is necessary for the client to maintain locally the set of unspent outputs for each viewing key, for each smart contract. This operation is called scanning. This local cache is updated downloading from the node only the difference from the last scanned state.

module Wallet : sig ... end